+++ src/Components/Studenthome.js
@@ -25,7 +25,7 @@ class Studenthome extends React.Component  {
         Authorization: 'bearer '+getTokens()
       },
     };
-    Axios.get('http://localhost:3001/dashboard/student',config)
+    Axios.get('http://localhost:3001/dashboard/',config)
     .then((res) => {
       console.log(res);
       this.setState({
